fix(ordering): fixes an issue when ordering field list is null
Fixes exception thrown when the ordering field list is null.

By default, ASP.NET Core parses collections as null when the property
query string is not present. I assumed it would parse it into
an empty collection.

This fix has a more sane default: it wont modify the IQueryable if
the list is null.